Ok avant precision

+++ libft/ft_strdup.c
@@ -6,21 +6,22 @@
 /*   By: cchampou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/06 01:23:12 by cchampou          #+#    #+#             */
-/*   Updated: 2016/11/22 11:23:58 by cchampou         ###   ########.fr       */
+/*   Updated: 2017/06/12 15:14:22 by cchampou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include "libft.h"
+#include <stdio.h>
 
 char	*ft_strdup(const char *s1)
 {
 	char	*output;
-	size_t	length;
-	size_t	cpt;
+	int		length;
+	int		cpt;
 
 	cpt = 0;
-	length = (size_t)ft_strlen(s1);
+	length = ft_strlen(s1);
 	if (!(output = (char*)malloc((length + 1) * sizeof(char))))
 		return (NULL);
 	while (cpt < length)
